requirements.txt
anyio==4.3.0
APScheduler==3.6.3
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
executing==2.0.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
ipython==8.22.2
jedi==0.19.1
matplotlib-inline==0.1.6
outcome==1.3.0.post0
parso==0.8.3
prompt-toolkit==3.0.43
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PySocks==1.7.1
python-dotenv==1.0.1
python-telegram-bot==13.5
pytz==2024.1
requests==2.31.0
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4
traitlets==5.14.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wcwidth==0.2.13
wsproto==1.2.0
